modutils-24: use xasprintf and xzalloc where appropriate,  remove unreachanble code build system: correct some dependencies in config system init: convert one #if into if() - for Rob.
function                                             old     new   delta
obj_string_patch                                     136     129      -7
bb_init_module_24                                   4759    4615    -144
------------------------------------------------------------------------------
(add/remove: 0/0 grow/shrink: 0/2 up/down: 0/-151)           Total: -151 bytes

+++ init/init.c
@@ -780,16 +780,16 @@ static void reload_inittab(void)
 		for (a = init_action_list; a; a = a->next)
 			if (a->pid != 0)
 				kill(a->pid, SIGTERM);
-#if CONFIG_FEATURE_KILL_DELAY
-		/* NB: parent will wait in NOMMU case */
-		if ((BB_MMU ? fork() : vfork()) == 0) { /* child */
-			sleep(CONFIG_FEATURE_KILL_DELAY);
-			for (a = init_action_list; a; a = a->next)
-				if (a->pid != 0)
-					kill(a->pid, SIGKILL);
-			_exit(EXIT_SUCCESS);
+		if (CONFIG_FEATURE_KILL_DELAY) {
+			/* NB: parent will wait in NOMMU case */
+			if ((BB_MMU ? fork() : vfork()) == 0) { /* child */
+				sleep(CONFIG_FEATURE_KILL_DELAY);
+				for (a = init_action_list; a; a = a->next)
+					if (a->pid != 0)
+						kill(a->pid, SIGKILL);
+				_exit(EXIT_SUCCESS);
+			}
 		}
-#endif
 	}
 
 	/* Remove old and unused entries */
